style: make each builtin a separate file
